+++ packages/ondevice-actions/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @storybook/addon-ondevice-actions
 
+## 10.3.0-next.0
+
 ## 10.2.3
 
 ## 10.2.2